feat: moves data sync components to namespaced directory
