+++ src/components/syncs/datasets/tables.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { useMutation } from "react-query";
 import { Button } from "@material-ui/core";
 import AccessDisplay from "./AccessDisplay";
